fix: compile error fixes
