+++ PrimerSDK.xcworkspace/contents.xcworkspacedata
@@ -216,9 +216,6 @@
                   </FileRef>
                </Group>
                <FileRef
-                  location = "group:SuccessMessage.swift">
-               </FileRef>
-               <FileRef
                   location = "group:_PaymentAPIModel.swift">
                </FileRef>
                <FileRef
@@ -360,9 +357,6 @@
                   location = "group:PrimerIntegrationOptions.swift">
                </FileRef>
                <FileRef
-                  location = "group:CheckoutModule.swift">
-               </FileRef>
-               <FileRef
                   location = "group:TokenizationRequestPaymentSessionInfo.swift">
                </FileRef>
                <FileRef
@@ -441,9 +435,6 @@
                   location = "group:UIScreenExtension.swift">
                </FileRef>
                <FileRef
-                  location = "group:WebViewUtil.swift">
-               </FileRef>
-               <FileRef
                   location = "group:NSObject+ClassName.swift">
                </FileRef>
                <FileRef
@@ -451,9 +442,6 @@
                </FileRef>
                <FileRef
                   location = "group:VersionUtils.swift">
-               </FileRef>
-               <FileRef
-                  location = "group:PresentationController.swift">
                </FileRef>
                <FileRef
                   location = "group:NSErrorExtension.swift">
@@ -480,9 +468,6 @@
                   location = "group:PrimerCancellable.swift">
                </FileRef>
                <FileRef
-                  location = "group:PostalCode.swift">
-               </FileRef>
-               <FileRef
                   location = "group:IntExtension.swift">
                </FileRef>
                <FileRef
@@ -495,12 +480,6 @@
                   location = "group:AlertController.swift">
                </FileRef>
                <FileRef
-                  location = "group:PrimerTableViewCell.swift">
-               </FileRef>
-               <FileRef
-                  location = "group:URLExtension.swift">
-               </FileRef>
-               <FileRef
                   location = "group:PrimerStackVIew.swift">
                </FileRef>
                <FileRef
@@ -508,9 +487,6 @@
                </FileRef>
                <FileRef
                   location = "group:PrimerImageView.swift">
-               </FileRef>
-               <FileRef
-                  location = "group:PrimerCustomStyleTextField.swift">
                </FileRef>
                <FileRef
                   location = "group:Mask.swift">
@@ -523,9 +499,6 @@
                </FileRef>
                <FileRef
                   location = "group:DeeplinkUtils.swift">
-               </FileRef>
-               <FileRef
-                  location = "group:PrimerScrollView.swift">
                </FileRef>
                <FileRef
                   location = "group:UIDeviceExtension.swift">
@@ -571,9 +544,6 @@
                </FileRef>
                <FileRef
                   location = "group:AnyEncodable.swift">
-               </FileRef>
-               <FileRef
-                  location = "group:Weak.swift">
                </FileRef>
             </Group>
             <Group
@@ -2218,9 +2188,6 @@
             <FileRef
                location = "group:CurrencyStorageTests.swift">
             </FileRef>
-            <FileRef
-               location = "group:PostalCodeTests.swift">
-            </FileRef>
          </Group>
          <Group
             location = "group:CVV Recapture"
@@ -2349,9 +2316,6 @@
             </FileRef>
             <FileRef
                location = "group:StringTests.swift">
-            </FileRef>
-            <FileRef
-               location = "group:WebViewUtilTests.swift">
             </FileRef>
             <FileRef
                location = "group:DateTests.swift">
@@ -2574,7 +2538,7 @@
                location = "group:Services"
                name = "Services">
                <FileRef
-                  location = "group:CreateResumePaymentService.swift">
+                  location = "group:MockCreateResumePaymentService.swift">
                </FileRef>
                <FileRef
                   location = "group:MockAPIClient.swift">
